Move and rename functions for easier discoverability
